Updated config files, replace config files per application, don't install applications that already exist on the machine (like docker and kodi)

+++ install.sh
@@ -1,22 +1,34 @@
 REPO_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 cd $REPO_DIR
 
+echo "enabled=0" | sudo tee /etc/default/apport > /dev/null
+
 source $REPO_DIR/config/globals.conf
 
-declare -A variables
-while IFS='= ' read var val
-do
-    if [ ! -z "$var" ]; then
-        variables["$var"]="$val"
-    fi
-done < config/globals.conf
-
-if [ -z "${variables['ROOT_FOLDER']}" ]; then
-    echo ""
-    echo "ERROR: You should configure it first. See file config/globals.conf"
-    echo ""
+if [ -z "${ROOT_FOLDER}" ]; then
+    echo "ERROR: You should configure ROOT_FOLDER first. See file config/globals.conf"
     exit
 fi
+if [ -z "${STORAGE_ROOT_FOLDER}" ]; then
+    echo "ERROR: You should configure STORAGE_ROOT_FOLDER first. See file config/globals.conf"
+    exit
+fi
+if [ -z "${IP_ADDRESS}" ]; then
+    echo "ERROR: You should configure IP_ADDRESS first. See file config/globals.conf"
+    exit
+fi
+
+
+
+# Setup download folders
+storageFolder=$ROOT_FOLDER/storage
+downloadsFolder=$storageFolder/downloads
+movies=$downloadsFolder/movies
+unordered=$downloadsFolder/unordered
+torrents=$downloadsFolder/torrents
+series=$downloadsFolder/series
+downloading=$downloadsFolder/downloading
+subtitles=$downloadsFolder/subtitles
 
 # Prepare root folder
 if [ ! -d $ROOT_FOLDER ]; then 
@@ -31,16 +43,6 @@ if [ -L storage ]; then
 fi
 ln -s "$STORAGE_ROOT_FOLDER" storage
 
-# Setup download folders
-storageFolder=$ROOT_FOLDER/storage
-downloadsFolder=$storageFolder/downloads
-movies=$downloadsFolder/movies
-unordered=$downloadsFolder/unordered
-torrents=$downloadsFolder/torrents
-series=$downloadsFolder/series
-downloading=$downloadsFolder/downloading
-subtitles=$downloadsFolder/subtitles
-
 # Create docker folder
 cd $storageFolder
 dockerStorageFolder=$storageFolder/.mediacenter-docker-files
@@ -50,6 +52,16 @@ fi
 
 
 
+declare -A variables
+variables['couchpotato.movies']="/transmission/download-movies"
+variables['couchpotato.password']="$USER"
+variables['couchpotato.IP_ADDRESS']="$IP_ADDRESS"
+variables['sickrage.series']="/transmission/download-series"
+variables['sickrage.IP_ADDRESS']="$IP_ADDRESS"
+variables['transmission.downloading']="/transmission/incomplete"
+variables['transmission.home']="/transmission"
+variables['transmission.unordered']="/transmission/download"
+
 
 rm -rf $REPO_DIR/config/generated
 mkdir $REPO_DIR/config/generated
@@ -57,17 +69,23 @@ cp -r $REPO_DIR/config-templates/* $REPO_DIR/config/generated
 
 cd $REPO_DIR
 for i in $REPO_DIR/config/generated/*; do
-    #echo "Processing config file $i"
+#    echo "Processing config file $i"
     cp ${i} tmp
     for j in ${!variables[@]}; do
-        result="sed -e 's|<${j}>|${variables[$j]}|g' tmp > tmp2"
-        #echo "Replacing var $j: $result"
+        result="sed -e 's|\${${j}}|${variables[$j]}|g' tmp > tmp2"
+#        echo "Replacing var $j: $result"
         eval $result
 
         mv tmp2 tmp
     done
     mv tmp "$i"
 done
+
+grep "\${" $REPO_DIR/config/generated/*
+if [ "$?" -eq 0 ]; then
+  echo "There are unresolved variables in the config files (see above lines). Exiting." 
+  exit
+fi
 
 
 #echo "Setting up series"
@@ -77,37 +95,77 @@ while IFS=$'    ' read -r f1 f2; do
 done < $REPO_DIR/config/series.csv
 
 
-
-sudo add-apt-repository -y ppa:team-xbmc/ppa
-
-sudo apt-get update
-sudo apt-get install linux-image-generic curl
-
-# Get the latest Docker package
-curl -sSL https://get.docker.com/ | sh
-
-# Create the docker group and add your user.
-sudo usermod -aG docker "${USER}"
+cp $REPO_DIR/config/kodi.desktop $HOME/.config/autostart/
+cp $REPO_DIR/config/generated/start-mediacenter.desktop $HOME/.config/autostart/
 
 
+if ! [ -x "$(command -v kodi)" ]; then
+  echo "Package kodi (xbmc) is not installed. Installing now"
+  sudo add-apt-repository -y ppa:team-xbmc/ppa >&2
+  sudo apt-get update
+  sudo apt-get install xbmc >&2
+fi
 
-docker rm -f `docker ps --no-trunc -aq`
+
+# Generate the kodi configuration
+cd $HOME/.kodi/userdata/
+if [ ! -d $dockerStorageFolder/kodi ]; then 
+  mkdir $dockerStorageFolder/kodi
+fi
+if [ ! -f $dockerStorageFolder/kodi/sources.xml ]; then 
+  cp $REPO_DIR/config/generated/kodi-sources.xml $dockerStorageFolder/kodi/sources.xml
+fi
+rm sources.xml
+ln -s $dockerStorageFolder/kodi/sources.xml
+
+if [ ! -f $dockerStorageFolder/kodi/guisettings.xml ]; then 
+  cp $REPO_DIR/config/generated/kodi-guisettings.xml $dockerStorageFolder/kodi/guisettings.xml
+fi
+rm guisettings.xml
+ln -s $dockerStorageFolder/kodi/guisettings.xml
 
 
 
-# Install File manager
-#docker build -t cron cron
+
+
+#if ! [ -x "$(command -v linux-image-generic)" ]; then
+#  echo "Package linux-image-generic is not installed. Installing now"
+#  sudo apt-get install linux-image-generic
+#fi
+if ! [ -x "$(command -v curl)" ]; then
+  echo "Package curl is not installed. Installing now"
+  sudo apt-get install curl
+fi
+
+
+
+if ! [ -x "$(command -v docker)" ]; then
+  # Get the latest Docker package
+  curl -sSL https://get.docker.com/ | sh
+
+  # Create the docker group and add your user.
+  sudo usermod -aG docker "${USER}"
+fi
+
+
+# Install the crontab
 crontab $REPO_DIR/cron/crontab
 
 
 
 
+# Delete all docker containers
+docker rm -f `docker ps --no-trunc -aq`
+
 #### Install transmission
-cd $REPO_DIR
-git clone https://github.com/dgholz/docker-transmission.git
-cd docker-transmission
-docker build -t transmission .
-rm -rf $REPO_DIR/docker-transmission
+docker ps -a | grep -q "mediacenter_transmission"
+if [ "$?" -eq 0 ]; then
+  cd $REPO_DIR
+  git clone https://github.com/dgholz/docker-transmission.git
+  cd docker-transmission
+  docker build -t transmission .
+  rm -rf $REPO_DIR/docker-transmission
+fi
 
 # Override config
 mkdir -p $dockerStorageFolder/transmission/config/
@@ -116,23 +174,57 @@ if [ ! -f $dockerStorageFolder/transmission/config/settings.json ]; then
 fi
 
 
+docker run -d \
+ --name mediacenter_transmission \
+ -v $downloadsFolder/series:/transmission/download-series \
+ -v $downloadsFolder/movies:/transmission/download-movies \
+ -v $downloadsFolder/downloading:/transmission/incomplete \
+ -v $downloadsFolder/unordered:/transmission/download \
+ -v $downloadsFolder/torrents:/transmission/watch \
+ -v $dockerStorageFolder/transmission/config:/transmission/config \
+ -p 9091:9091 \
+ -p 51413:51413 \
+ -p 51413:51413/udp \
+ --restart=always \
+ transmission
 
+docker run -d \
+ --name mediacenter_transmissionClient \
+ --link mediacenter_transmission:transmission \
+ --restart=always \
+ busybox
 
 
 
 
 #### Install Sickbeard
 cd $REPO_DIR
-git clone http://github.com/timhaak/docker-sickrage.git
-cd docker-sickrage
-docker build -t sickrage .
-rm -rf $REPO_DIR/docker-sickrage
+docker ps -a | grep -q "mediacenter_sickrage"
+if [ "$?" -eq 0 ]; then
+  git clone http://github.com/timhaak/docker-sickrage.git
+  cd docker-sickrage
+  docker build -t sickrage .
+  rm -rf $REPO_DIR/docker-sickrage
+fi
 
 # Override config
 mkdir -p $dockerStorageFolder/sickrage/config/
 if [ ! -f $dockerStorageFolder/sickrage/config/config.ini ]; then 
   cp $REPO_DIR/config/generated/sickbeard.ini $dockerStorageFolder/sickrage/config/config.ini
 fi
+
+docker run -d \
+ --name mediacenter_sickrage \
+ -h localhost \
+ -v $dockerStorageFolder/sickrage/config:/config \
+ -v $downloadsFolder/series:/transmission/download-series \
+ -p 8081:8081 \
+ --restart=always \
+ sickrage
+
+
+
+
 
 
 
@@ -142,11 +234,14 @@ fi
 
 
 #### Install couchpotato
-cd $REPO_DIR
-git clone https://github.com/timhaak/docker-couchpotato.git
-cd docker-couchpotato
-docker build -t couchpotato .
-rm -rf $REPO_DIR/docker-couchpotato
+docker ps -a | grep -q "mediacenter_couchpotato"
+if [ "$?" -eq 0 ]; then
+  cd $REPO_DIR
+  git clone https://github.com/timhaak/docker-couchpotato.git
+  cd docker-couchpotato
+  docker build -t couchpotato .
+  rm -rf $REPO_DIR/docker-couchpotato
+fi
 
 # Override config
 mkdir -p $dockerStorageFolder/couchpotato/config/
@@ -154,64 +249,15 @@ if [ ! -f $dockerStorageFolder/couchpotato/config/CouchPotato.cfg ]; then
   cp $REPO_DIR/config/generated/couchpotato.cfg $dockerStorageFolder/couchpotato/config/CouchPotato.cfg
 fi
 
-
-
-
-#docker run -d \
-# --name cron \
-# -v $dockerStorageFolder/data/downloads:/data \
-# -v $dockerStorageFolder/data/shazam-tags:/shazam-tags \
-# cron
-
 docker run -d \
- --name transmission1 \
- -v $downloadsFolder/unordered:/transmission/download \
- -v $dockerStorageFolder/transmission/config:/transmission/config \
- -p 9091:9091 \
- -p 51413:51413 \
- -p 51413:51413/udp \
- transmission
-
-docker run \
- --name transmissionClient \
- --link transmission1:transmission \
- busybox
-
-docker run -d \
- --name sickrage \
- -h localhost \
- -v $dockerStorageFolder/sickrage/config:/config \
- -v $downloadsFolder/series:/data \
- -p 8081:8081 \
- sickrage
-
-
-# Run docker container
-docker run -d \
- --name couchpotato \
+ --name mediacenter_couchpotato \
  -h localhost \
  -v $dockerStorageFolder/couchpotato/config:/config \
- -v $downloadsFolder/movies:/data \
+ -v $downloadsFolder/movies:/transmission/download-movies \
  -p 5050:5050 \
+ --restart=always \
  couchpotato
 
 
-cp $REPO_DIR/config/kodi.desktop $HOME/.config/autostart/
-cp $REPO_DIR/config/generated/start-mediacenter.desktop $HOME/.config/autostart/
 
-echo "enabled=0" | sudo tee /etc/default/apport
 
-#### Install KODI
-echo "Installing XBMC"
-sudo apt-get install xbmc
-
-# Generate the kodi configuration
-if [ ! -d $dockerStorageFolder/kodi ]; then 
-  mkdir $dockerStorageFolder/kodi
-fi
-if [ ! -f $dockerStorageFolder/kodi/sources.xml ]; then 
-  cp $REPO_DIR/config-templates/kodi-sources.xml $dockerStorageFolder/kodi/sources.xml
-fi
-cd $HOME/.kodi/userdata/
-rm sources.xml
-ln -s $dockerStorageFolder/kodi/sources.xml
